pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vs2"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "pandas",
    "pyarrow",
    "pydantic",
    "rapidfuzz",
    "python-dateutil",
    "pyyaml",
    "tqdm",
]

[tool.setuptools.packages.find]
where = ["src"]
